Renamed Graph to Tape, moved Tape, Var, Param to autograd, removed unused imports, bugfixed.
