+++ src/main.rs
@@ -1,5 +1,3 @@
-use autograph as ag;
-
 fn main() {
-  println!("{:e}", 10f32.exp()); 
+  
 }
